// src/app/models/product.ts
export class product {
    productId: string
    productName: string
    brand: string
    model: string
    type: string
    costNew: number
    costAvg: number
    price1: number
    price2: number
    price3: number
    counterProduct: number
    balance: number
    status: string
    userUpdate: string
    updationDatetime: Date
    creationDatetime: Date
    // imgPath: string
}